🚧 (wip) extract state from slope charts
