+++ packages/@ourworldindata/grapher/src/slopeCharts/SlopeChartConstants.ts
@@ -6,6 +6,13 @@ import {
 } from "@ourworldindata/types"
 import { ChartSeries } from "../chart/ChartInterface"
 import { CoreColumn } from "@ourworldindata/core-table"
+import { ChartManager } from "../chart/ChartManager"
+
+export interface SlopeChartManager extends ChartManager {
+    canSelectMultipleEntities?: boolean // used to pick an appropriate series name
+    hasTimeline?: boolean // used to filter the table for the entity selector
+    hideNoDataSection?: boolean
+}
 
 export interface SlopeChartSeries extends ChartSeries {
     column: CoreColumn
